Translate documentation from SGML to RST

+++ Makefile
@@ -5,10 +5,12 @@ default:
 include Version
 
 tar:
+	@if [ `which autoconf` = '' ] ; then echo "Please install 'autoconf'" ; exit 1 ; fi
+	@if [ `which db2html` = '' ] ; then echo "Please install 'db2html' package 'docbook', 'docbook-utils', 'docbook-to-man', 'docbook2x', 'sgml-tool' )" ; exit 1 ; fi
 	@echo "The tar will be created, untarred, compiled and tested"
 	@echo "press return"
 	@read A
-	-rm *~
+	-rm *~ 2>/dev/null
 	cd src ; $(MAKE) clean ; autoconf ; rm -fR Makefile auto*.cache || true
 	cd tst ; $(MAKE) clean
 	cd doc ; $(MAKE) clean nnfs.html
